Move BfDs component examples to separate __examples__ files
Improve code organization by extracting component examples from main component
files into dedicated example files. This separates demonstration code from
core component logic, making components cleaner and examples more discoverable.

Changes:
- Create apps/bfDs/components/__examples__/ folder with 14 example files
- Extract .Example functions from all BfDs components to separate .example.tsx files
- Remove example code and unused imports from main component files
- Update Demo.tsx to import examples from new __examples__ location
- Maintain all original example functionality and styling

Test plan:
1. Run build: `bft build`
2. Start development server: `bft devTools`
3. Navigate to BfDs demo page
4. Verify all component examples still work correctly
5. Check that component files are cleaner without example code
