+++ apps/bfDs/components/__examples__/BfDsCallout.example.tsx
@@ -1,0 +1,160 @@
+import * as React from "react";
+import { BfDsCallout, type BfDsCalloutVariant } from "../BfDsCallout.tsx";
+
+export function BfDsCalloutExample() {
+  const [notifications, setNotifications] = React.useState<
+    Array<{
+      id: number;
+      message: string;
+      variant: BfDsCalloutVariant;
+      details?: string;
+    }>
+  >([]);
+  const [nextId, setNextId] = React.useState(1);
+
+  const addNotification = (
+    message: string,
+    variant: BfDsCalloutVariant,
+    details?: string,
+  ) => {
+    setNotifications(
+      (prev) => [...prev, { id: nextId, message, variant, details }],
+    );
+    setNextId((prev) => prev + 1);
+  };
+
+  const removeNotification = (id: number) => {
+    setNotifications((prev) => prev.filter((n) => n.id !== id));
+  };
+
+  const sampleData = {
+    name: "John Doe",
+    email: "john@example.com",
+    preferences: {
+      theme: "dark",
+      notifications: true,
+    },
+  };
+
+  return (
+    <div
+      style={{
+        display: "flex",
+        flexDirection: "column",
+        gap: "24px",
+        padding: "24px",
+        backgroundColor: "var(--bfds-background)",
+        color: "var(--bfds-text)",
+        fontFamily: "system-ui, -apple-system, sans-serif",
+        maxWidth: "600px",
+      }}
+    >
+      <h2>BfDsCallout Examples</h2>
+
+      <div>
+        <h3>Static Examples</h3>
+        <div style={{ display: "flex", flexDirection: "column", gap: "16px" }}>
+          <BfDsCallout
+            message="This is an informational message"
+            variant="info"
+          />
+          <BfDsCallout
+            message="Operation completed successfully!"
+            variant="success"
+          />
+          <BfDsCallout
+            message="Please review your settings"
+            variant="warning"
+          />
+          <BfDsCallout
+            message="An error occurred while processing"
+            variant="error"
+          />
+        </div>
+      </div>
+
+      <div>
+        <h3>With Details</h3>
+        <BfDsCallout
+          message="Form submitted successfully"
+          variant="success"
+          details={JSON.stringify(sampleData, null, 2)}
+          onDismiss={() => {}}
+        />
+      </div>
+
+      <div>
+        <h3>Dynamic Notifications</h3>
+        <div
+          style={{
+            display: "flex",
+            gap: "12px",
+            flexWrap: "wrap",
+            marginBottom: "16px",
+          }}
+        >
+          <button
+            type="button"
+            onClick={() => addNotification("Info notification", "info")}
+            style={{
+              padding: "8px 16px",
+              backgroundColor: "var(--bfds-primary)",
+              color: "var(--bfds-background)",
+              border: "none",
+              borderRadius: "4px",
+              cursor: "pointer",
+            }}
+          >
+            Add Info
+          </button>
+          <button
+            type="button"
+            onClick={() => addNotification("Success!", "success")}
+            style={{
+              padding: "8px 16px",
+              backgroundColor: "var(--bfds-success)",
+              color: "var(--bfds-background)",
+              border: "none",
+              borderRadius: "4px",
+              cursor: "pointer",
+            }}
+          >
+            Add Success
+          </button>
+          <button
+            type="button"
+            onClick={() =>
+              addNotification(
+                "Form data saved",
+                "success",
+                JSON.stringify(sampleData, null, 2),
+              )}
+            style={{
+              padding: "8px 16px",
+              backgroundColor: "var(--bfds-success)",
+              color: "var(--bfds-background)",
+              border: "none",
+              borderRadius: "4px",
+              cursor: "pointer",
+            }}
+          >
+            Add with Details
+          </button>
+        </div>
+
+        <div style={{ display: "flex", flexDirection: "column", gap: "12px" }}>
+          {notifications.map((notification) => (
+            <BfDsCallout
+              key={notification.id}
+              message={notification.message}
+              variant={notification.variant}
+              details={notification.details}
+              onDismiss={() => removeNotification(notification.id)}
+              autoDismiss={5000}
+            />
+          ))}
+        </div>
+      </div>
+    </div>
+  );
+}
